+++ FE/user_pages/src/components/bookDetail/InfoTable.jsx
