+++ FE/user_pages/src/components/home/SearchBar.jsx
